+++ mayan/apps/storage/backends/compressedstorage.py
@@ -96,7 +96,7 @@ class ZipCompressedPassthroughStorage(PassthroughStorage):
             ):
                 name = self._call_backend_method(
                     method_name='save', kwargs={
-                        'content': ContentFile(content=''), 'name': name
+                        'content': ContentFile(content=b''), 'name': name
                     }
                 )